+++ lib/postgres/table.d.ts
@@ -1,9 +1,9 @@
-import { Table, TablePartial, Value, Datum, SchemaEntry, DeepPartial, WriteResult, SingleSelection, Selection } from '../types';
+import { Table, TablePartial, IndexChangeResult, Value, Datum, SchemaEntry, DeepPartial, WriteResult, SingleSelection, Selection } from '../types';
 import { WrappedPostgresDatabase } from './wrapper';
 import { PostgresStream } from './stream';
 export declare class PostgresTablePartial<T = any> extends PostgresStream<T> implements TablePartial<T> {
     private types;
-    private readonly primaryIndexGetter;
+    private get primaryIndexGetter();
     constructor(db: WrappedPostgresDatabase, tableName: Value<string>, types: Value<SchemaEntry[]>);
     filter(predicate: DeepPartial<T> | ((doc: Datum<T>) => Value<boolean>)): Selection<T>;
     fork(): never;
@@ -22,6 +22,11 @@ export declare class PostgresTablePartial<T = any> extends PostgresStream<T> imp
     insert(obj: T, options?: {
         conflict: 'error' | 'replace' | 'update';
     }): Datum<WriteResult<T>>;
+    indexCreate<U extends keyof T>(key: U): Datum<IndexChangeResult>;
+    indexCreate(key: any): Datum<IndexChangeResult>;
+    indexDrop<U extends keyof T>(key: U): Datum<IndexChangeResult>;
+    indexDrop(key: any): Datum<IndexChangeResult>;
+    indexList(): Datum<any[]>;
     run(): Promise<T[]>;
 }
 interface PostgresTable<T = any> extends PostgresTablePartial<T>, Table<T> {