+++ src/postgres/selection.ts
@@ -46,12 +46,12 @@ export class PostgresSelectionPartial<T = any> extends PostgresStream<T> impleme
 
         const poost = (post ? ' AND ' + post : '') + (limit ?  ' LIMIT ' + limit : '');
         return this.db.get<{
-          'COUNT(*)': number
+          'count': number
         }>(`SELECT COUNT(*) FROM ${JSON.stringify(tableName)} WHERE ${selection}${poost}`)
-          .then(a => limit ? Math.min(a['COUNT(*)'], limit) : a['COUNT(*)']);
+          .then(a => limit ? Math.min(a['count'], limit) : a['count']);
       }
-      return this.db.get<{ 'COUNT(*)': number }>
-        (`SELECT COUNT(*) FROM ${JSON.stringify(tableName)} WHERE ${selection}`).then(a => a['COUNT(*)']);
+      return this.db.get<{ 'count': number }>
+        (`SELECT COUNT(*) FROM ${JSON.stringify(tableName)} WHERE ${selection}`).then(a => a['count']);
     }));
   }
 
@@ -147,6 +147,5 @@ export function createSelection<T = any>(db: WrappedPostgresDatabase, tableName:
       run() { return instance.run(); }
     } as SelectionPartial<T>) as any;
   (o as any).__proto__ = instance;
-  return o as any;
   return o as any;
 }