+++ src/postgres/table.ts
@@ -1,5 +1,5 @@
 import {
-  Table, TablePartial,
+  Table, TablePartial, IndexChangeResult,
   Value, Datum, SchemaEntry, DeepPartial, WriteResult,
   SingleSelection, Selection, Query, Stream
 } from '../types';
@@ -41,10 +41,10 @@ export class PostgresTablePartial<T = any> extends PostgresStream<T> implements 
         this.query = [];
 
         const poost = (post ? ` WHERE ${post}` : '') + (limit ? `LIMIT ${limit}` : '');
-        return this.db.get<{ 'COUNT(*)': number }>(`SELECT COUNT(*) FROM ${JSON.stringify(tableName)}${poost}`)
-          .then(a => limit ? Math.min(a['COUNT(*)'], limit) : a['COUNT(*)']);
+        return this.db.get<{ 'count': number }>(`SELECT COUNT(*) FROM ${JSON.stringify(tableName)}${poost}`)
+          .then(a => limit ? Math.min(a['count'], limit) : a as any);
       }
-      return this.db.get<{ 'COUNT(*)': number }>(`SELECT COUNT(*) FROM ${JSON.stringify(tableName)}`).then(a => a['COUNT(*)']);
+      return this.db.get<{ 'count': number }>(`SELECT COUNT(*) FROM ${JSON.stringify(tableName)}`).then(a => a['count']);
     }));
   }
 
@@ -125,6 +125,37 @@ export class PostgresTablePartial<T = any> extends PostgresStream<T> implements 
         () => ({ deleted: 0, skipped: 0, errors: 0, inserted: 1, replaced: 0, unchanged: 0 }),
         e => ({ deleted: 0, skipped: 1, errors: 1, first_error: String(e), inserted: 0, replaced: 0, unchanged: 1 }));
       return ret;
+    }));
+  }
+
+  indexCreate<U extends keyof T>(key: U): Datum<IndexChangeResult>;
+  indexCreate(key: any): Datum<IndexChangeResult>;
+  // indexCreate(name: Value<String>, indexFunction: (doc: Datum<T>) => Value<boolean>): Datum<IndexChangeResult>;
+  indexCreate(key: any): Datum<IndexChangeResult> {
+    return expr(createQuery(async () => {
+      const tableName = await resolveValue(this.tableName);
+      await this.db.exec(`CREATE INDEX ${JSON.stringify(tableName + '_' + key)} ON ${JSON.stringify(tableName)}(${JSON.stringify(key)})`);
+      return { created: 1 } as IndexChangeResult;
+    }));
+  }
+
+  indexDrop<U extends keyof T>(key: U): Datum<IndexChangeResult>;
+  // indexDrop(name: Value<string>): Datum<IndexChangeResult>;
+  indexDrop(key: any): Datum<IndexChangeResult>;
+  indexDrop(key: any): Datum<IndexChangeResult> {
+    return expr(createQuery(async () => {
+      const tableName = await resolveValue(this.tableName);
+      await this.db.exec(`DROP INDEX ${JSON.stringify(tableName + '_' + key)}`);
+      return { dropped: 1 } as IndexChangeResult;
+    }));
+  }
+
+  indexList(): Datum<any[]> {
+    return expr(createQuery(async () => {
+      const tableName = await resolveValue(this.tableName);
+      const rows = await this.db.all<{ indexname: string }>(
+        `SELECT indexname FROM pg_indexes WHERE tablename=${safen(tableName)};`);
+      return rows.map(row => row.indexname.slice(tableName.length + 1));
     }));
   }